dom/tests/mochitest/general/test_contentViewer_overrideDPPX.html

<!DOCTYPE HTML>
<html>
<head>
  <title>nsIContentViewer::overrideDPPX test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
</head>

<body>

<p id="display"></p>

<iframe></iframe>

<script type="application/javascript;version=1.8">

SimpleTest.waitForExplicitFinish();

const frameWindow = document.querySelector("iframe").contentWindow;

const docShellFor = (win) => 
  SpecialPowers.wrap(win)
    .QueryInterface(SpecialPowers.Ci.nsIInterfaceRequestor)
    .getInterface(SpecialPowers.Ci.nsIWebNavigation)
    .QueryInterface(SpecialPowers.Ci.nsIDocShell);

const docShell = docShellFor(window);
const frameDocShell = docShellFor(frameWindow);

const originalDPR = window.devicePixelRatio;
const originalZoom = docShell.contentViewer.fullZoom;
const dppx = originalDPR * 1.5;
const zoom = originalZoom * 0.5;

const overrideDPPX = (value) => {
  if (value > 0) {
    info(`override window's dppx to ${value}`);
  } else {
    info(`restore window's dppx to default value`);
  }

   docShell.contentViewer.overrideDPPX = value;
}

const fullZoom = (value) => {
  info(`set window's fullZoom to ${value}`);
  docShell.contentViewer.fullZoom = value;
}

const createFontStyleForDPPX = (doc, value, size) => {
  info(`adding a stylesheet that set font size to ${size}px for ${value}dppx`);

  let style = doc.createElement("style");

  style.setAttribute("media", `(resolution: ${value}dppx)`);

  doc.head.appendChild(style);

  style.sheet.insertRule(`body {font-size: ${size}px}`, 0);

  return style;
}

const ensureIntegrityInitialValues = () => {
  is(window.devicePixelRatio, originalDPR,
    "devicePixelRatio has the original value.");
  is(docShell.contentViewer.fullZoom, originalZoom,
    "fullZoom has the original value.");

  is(frameWindow.devicePixelRatio, originalDPR,
    "devicePixelRatio has the original value.");
  is(frameDocShell.contentViewer.fullZoom, originalZoom,
    "fullZoom has the original value.");
}

const gTests = {
  "test overrideDPPX with devicePixelRatio": (done) => {
    ensureIntegrityInitialValues();

    overrideDPPX(dppx);

    is(window.devicePixelRatio, dppx,
      "devicePixelRatio overridden.");
    is(frameWindow.devicePixelRatio, dppx,
      "frame's devicePixelRatio overridden.");

    overrideDPPX(0);

    is(window.devicePixelRatio, originalDPR,
      "devicePixelRatio back to default.");
    is(frameWindow.devicePixelRatio, originalDPR,
      "frame's devicePixelRatio back to default.");

    done();
  },
  "test overrideDPPX with devicePixelRatio and fullZoom": (done) => {
    ensureIntegrityInitialValues();

    fullZoom(zoom);
    overrideDPPX(dppx);

    is(window.devicePixelRatio, dppx,
      "devicePixelRatio overridden, fullZoom ignored");
    is(frameWindow.devicePixelRatio, dppx,
      "frame's devicePixelRatio overridden, fullZoom ignored");

    overrideDPPX(0);

    is(window.devicePixelRatio, originalDPR * zoom,
      "devicePixelRatio now is affected by fullZoom");
    is(frameWindow.devicePixelRatio, originalDPR * zoom,
      "frame's devicePixelRatio now is affected by fullZoom");

    fullZoom(originalZoom);

    is(window.devicePixelRatio, originalDPR,
      "devicePixelRatio back to default.");
    is(frameWindow.devicePixelRatio, originalDPR,
      "frame's devicePixelRatio back to default.");

    done();

  },
  "test overrideDPPX with media queries": (done) => {
    ensureIntegrityInitialValues();

    let frameDoc = frameWindow.document;

    let originalFontSize = getComputedStyle(document.body).fontSize;
    let frameOriginalFontSize = getComputedStyle(frameDoc.body).fontSize;

    let style = createFontStyleForDPPX(document, dppx, "32");
    let frameStyle = createFontStyleForDPPX(frameDoc, dppx, "32");

    let currentFontSize = getComputedStyle(document.body).fontSize;
    let frameCurrentFontSize = getComputedStyle(frameDoc.body).fontSize;

    is(currentFontSize, originalFontSize,
      "media query are not applied yet");
    is(frameCurrentFontSize, frameOriginalFontSize,
      "frame's media query are not applied yet");

    overrideDPPX(dppx);

    currentFontSize = getComputedStyle(document.body).fontSize;
    frameCurrentFontSize = getComputedStyle(frameDoc.body).fontSize;

    isnot(currentFontSize, originalFontSize,
      "media query are applied.");
    isnot(frameCurrentFontSize, frameOriginalFontSize,
      "frame's media query are applied.");

    is(currentFontSize, "32px",
      "font size has the expected value.");
    is(frameCurrentFontSize, "32px",
      "frame's font size has the expected value.");

    overrideDPPX(0);

    currentFontSize = getComputedStyle(document.body).fontSize;
    frameCurrentFontSize = getComputedStyle(frameDoc.body).fontSize;

    is(currentFontSize, originalFontSize,
      "media query are not applied anymore.");
    is(frameCurrentFontSize, frameOriginalFontSize,
      "media query are not applied anymore.");

    style.remove();
    frameStyle.remove();

    done();
  },
  "test overrideDPPX with media queries and fullZoom": (done) => {
    ensureIntegrityInitialValues();

    let frameDoc = frameWindow.document;

    let styles = [
      createFontStyleForDPPX(document, originalDPR, "16"),
      createFontStyleForDPPX(document, dppx, "32"),
      createFontStyleForDPPX(document, originalDPR * zoom, "48"),
      createFontStyleForDPPX(frameDoc, originalDPR, "16"),
      createFontStyleForDPPX(frameDoc, dppx, "32"),
      createFontStyleForDPPX(frameDoc, originalDPR * zoom, "48")
    ];

    let currentFontSize = getComputedStyle(document.body).fontSize;
    let frameCurrentFontSize = getComputedStyle(frameDoc.body).fontSize;
    is(currentFontSize, "16px",
      "media query are not applied yet");
    is(frameCurrentFontSize, "16px",
      "frame's media query are not applied yet");

    fullZoom(zoom);
    overrideDPPX(dppx);

    currentFontSize = getComputedStyle(document.body).fontSize;
    frameCurrentFontSize = getComputedStyle(frameDoc.body).fontSize;
    is(currentFontSize, "32px",
      "media query are applied for overridden DDPX, fullZoom ignored.");
    is(frameCurrentFontSize, "32px",
      "frame's media query are applied for overridden DDPX, fullZoom ignored.");

    overrideDPPX(0);

    currentFontSize = getComputedStyle(document.body).fontSize;
    frameCurrentFontSize = getComputedStyle(frameDoc.body).fontSize;
    is(currentFontSize, "48px",
      "media query are applied for fullZoom.");
    is(frameCurrentFontSize, "48px",
      "frame's media query are applied for fullZoom.");

    fullZoom(originalZoom);

    currentFontSize = getComputedStyle(document.body).fontSize;
    frameCurrentFontSize = getComputedStyle(frameDoc.body).fontSize;
    is(currentFontSize, "16px",
      "media query are not applied anymore.");
    is(frameCurrentFontSize, "16px",
      "frame's media query are not applied anymore.");

    styles.forEach(style => style.remove());

    done();
  },
  "test OverrideDPPX with MediaQueryList": (done) => {
    ensureIntegrityInitialValues();

    let promises = [
      new Promise(resolve => {
        let mql = window.matchMedia(`(resolution: ${dppx}dppx)`);

        mql.addListener(function listener() {
          ok("MediaQueryList's listener invoked.")
          mql.removeListener(listener);
          resolve();
        });
      }),
      new Promise(resolve => {
        let mql = frameWindow.matchMedia(`(resolution: ${dppx}dppx)`);

        mql.addListener(function listener() {
          ok("frame's MediaQueryList's listener invoked.")
          mql.removeListener(listener);
          resolve();
        });
      })
    ];

    Promise.all(promises)
      .then(() => overrideDPPX(0))
      .then(done, e => {throw e});

    overrideDPPX(dppx);
  },
  "test OverrideDPPX with MediaQueryList and fullZoom": (done) => {
    ensureIntegrityInitialValues();

    let mql1 = window.matchMedia(`(resolution: ${dppx}dppx)`);
    let mql2 = window.matchMedia(`(resolution: ${originalDPR * zoom}dppx)`);

    let overridden = false;

    mql1.addListener(function listener() {
      ok("MediaQueryList's listener for dppx invoked.");
      mql1.removeListener(listener);
      overridden = true;
      overrideDPPX(0);
    });

    mql2.addListener(function listener() {
      ok("MediaQueryList's listener for zoom invoked.");
      mql2.removeListener(listener);

      fullZoom(originalZoom);

      ok(overridden, "override DPPX and fullZoom applied in the proper order.");

      done();
    });

    overrideDPPX(dppx);
    fullZoom(zoom);
  }
};

function runner(tests) {
  for (let name of Object.keys(tests)) {
    info(name);
    tests[name](next);
    yield undefined;
  }

  SimpleTest.finish();
};

const gTestRunner = runner(gTests);

function next() {
  SimpleTest.executeSoon(function() {
    try {
      gTestRunner.next();
    } catch (e) {
      if (e !== StopIteration) { throw e; }
    }
  });
}

// Run the tests
addLoadEvent(next);

</script>

</body>
</html>
